[FIX, TEST] Remove individual split bin checks in create_ibfs_from_chopper_pack_test
Since the naive splitting is done with an std::unordered_set the hashing
and the order of kmers in the set is non-determinisitic accross
different platforms, binaries etc so we cannot safely test them. Testing
that the sum of the split bins has to add up to the expected total
amount has to be enough.